requirements.txt
## install dependencies:
#    pip install -r requirements
## check outdated dependencies:
#    pip list --outdated

Django==1.8.14
pytz
psycopg2
django-pipeline==1.6.8
django-autoslug==1.9.3
awesome-slugify==1.6.5
django-parler==1.5.1
django-ckeditor==5.0.3
django-taggit==0.17.6
django-taggit-autosuggest==0.2.8
django-select-multiple-field==0.4.2
django-el-pagination==2.1.2

pillow==3.2.0
sorl-thumbnail==12.3
WeasyPrint
html5lib==0.9999999  # the latest html5lib==0.999999999 install fails

redis==2.10.5
celery==3.1.23
Whoosh==2.7.4

django-simplesearch==1.0
unicodecsv
## astroedu
bleach==1.4.2
mistune==0.7.2
rtfunicode==1.0
